requirements.txt
python-telegram-bot[ext]
pyyaml
python-dotenv
pytz
